Moved test cases to service/simplify
