+++ test/za/ac/sun/cs/green/EntireSuite.java
@@ -32,7 +32,7 @@ import za.ac.sun.cs.green.service.z3.SATZ3Test;
 import za.ac.sun.cs.green.util.ParallelSATTest;
 import za.ac.sun.cs.green.util.SetServiceTest;
 import za.ac.sun.cs.green.util.SetTaskManagerTest;
-import za.ac.sun.cs.green.service.canonizer.OnlyConstantPropogationTest;
+import za.ac.sun.cs.green.service.simplify.OnlyConstantPropogationTest;
 
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
