+++ test/za/ac/sun/cs/green/service/simplify/OnlyConstantPropogationTest.java
