+++ test/za/ac/sun/cs/green/service/simplify/SimplificationConstantPropogationTest.java
